backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
# aioredis==2.0.1  # Removed - no Redis usage
aiosignal==1.3.2
alembic==1.12.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
Authlib==1.6.0
azure-core==1.34.0
azure-functions==1.23.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
bandit==1.8.3
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
celery==5.3.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.2
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
dparse==0.6.4
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.104.1
filelock==3.16.1
flake8==7.2.0
frozenlist==1.6.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kombu==5.5.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.4
mypy==1.16.0
mypy_extensions==1.1.0
nltk==3.9.1
openai==1.3.7
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
packaging==24.2
passlib[bcrypt]==1.7.4
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.5
psutil==6.1.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.23.4
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.31.0
rich==14.0.0
rsa==4.9.1
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
safety==3.5.2
safety-schemas==0.0.14
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.46.2
stevedore==5.4.1
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn[standard]==0.24.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yamllint==1.37.1
yarl==1.20.0
zipp==3.23.0
google-generativeai==0.3.2
anthropic==0.7.8
opentelemetry-auto-instrumentation==0.42b0
slowapi==0.1.9
limits==4.2
deprecated==1.2.18
